work project(student model)

+++ src/main/java/com/web_project/school/service/StudentService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface StudentService {
     public List<StudentModel> findAllStudents();
 
-    public StudentModel findStudentById(int id);
+    public StudentModel findStudentById(Long id);
 
     public StudentModel addStudent(StudentModel student);
 
     public StudentModel updateStudent(StudentModel student);
 
-    public void deleteStudent(int id);
+    public void deleteStudent(Long id);
 }
